backend/requirements.txt
# PDF to DOCX Converter - Python dependencies
# Install: pip install -r requirements.txt

# Core PDF processing
PyMuPDF>=1.24.0  # fitz

# DOCX generation
python-docx>=1.1.0

# Table extraction (optional but recommended)
pdfplumber>=0.10.0

# Pillow (images)
Pillow>=10.0.0

# FastAPI (web server, optional)
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9  # for file uploads
